kvserver: add snapshot_prepare.go
We'll extract all the engine interactions from `(*Replica).applySnapshot` into this file.

The intention is that after that point, we will extract into a package and add
unit testing.

Epic: CRDB-46488
Release note: none
